+++ zaifbot/modules/api/last_price.py
@@ -1,10 +1,12 @@
 import traceback
 from datetime import datetime, timedelta
 from threading import Thread, Event, Lock
+
 from zaifapi.impl import ZaifPublicStreamApi
-from zaifbot.bot_common.api.wrapper import BotPublicApi
+
 from zaifbot.bot_common.errors import ZaifBotError
 from zaifbot.bot_common.logger import logger
+from zaifbot.modules.api.wrapper import BotPublicApi
 from .cache import ZaifCurrencyPairs
